Just a few issues fixed to restore functionality
Although it's a far cry from testing on the testbed online, still helps
to test misc. apps\bot scripts on the local drive and it takes only a
few minutes to narrow down the issue of a script failing. Just trying to
keep this project from dying, it's a helpful tool in troubleshooting.